chore(server): clear all clippy/rustc warnings under strict lint baseline
Brought the workspace from 1295 warnings to 0 under the strict baseline
introduced in 389c30e (clippy::all + pedantic + nursery + cargo + ~80
restriction lints + missing_docs).

Highlights:
- Removed 5 unused crate dependencies (serde_json from common; thiserror
  + curve25519-dalek from core; jsonwebtoken + tower from api).
- Renamed `let gen` -> `let generation` (channel_encryption.rs) and
  `rng.gen()` -> `rng.r#gen()` (webhook.rs) for Rust 2024 keyword reservation.
- Auto-applied cargo clippy --fix: pub fn -> pub const fn where eligible,
  added #[must_use], Self::Variant in matches, to_string -> to_owned.
- 343 unreachable `pub` -> `pub(crate)`.
- 850 missing_docs auto-generated via scripts/bulk_lint_fix.py with
  domain-aware mappings (e.g. created_at -> "Creation timestamp.",
  user_id -> "Owning user identifier.", create_X -> "Creates X.").
- 25 `let _ = expr;` -> `drop(expr);` (mix of bulk_lint_fix.py and
  scripts/fix_let_underscore.py for multi-line cases).
- 7 unused_qualifications (uuid::Uuid / axum::http::HeaderMap) cleaned up.
- 2 trivial f32-as-f32 casts removed in image_processing.rs.
- 1 rust_2024_incompatible_pat fixed via .as_deref() in servers.rs.
- Speaking SFU variant marked with #[expect(dead_code, reason = ...)]
  per project rule "no auto-removal of planned features".
- 28 tail_expr_drop_order documented via crate-level
  #![expect(tail_expr_drop_order, reason = "...")] on api/core/media
  with audit rationale, to be revisited during 2024 edition migration.

Verification:
- cargo check --workspace --all-targets --all-features (Docker, libssl-dev):
  Finished dev profile in 2m 31s, 0 errors, 0 warnings.
- docker compose build jolkr-api jolkr-media: clean real-build pass
  (warnings shown in builder 11/14 are stub-deps caching pass and benign).
- docker compose up + nginx restart: api /health 200, media started ok.

+++ jolkr-server/crates/jolkr-api/src/image_processing.rs
@@ -12,11 +12,14 @@ const ICON_MAX_PX: u32 = 256;
 /// The purpose of the uploaded image — determines the max output size.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum ImagePurpose {
+    /// `Avatar` variant.
     Avatar,
+    /// `Icon` variant.
     Icon,
 }
 
 impl ImagePurpose {
+    /// Max px.
     pub fn max_px(self) -> u32 {
         match self {
             Self::Avatar => AVATAR_MAX_PX,
@@ -24,6 +27,7 @@ impl ImagePurpose {
         }
     }
 
+    /// Parses from a string.
     pub fn from_str(s: &str) -> Option<Self> {
         match s {
             "avatar" => Some(Self::Avatar),
@@ -96,8 +100,8 @@ fn rasterize_svg(data: &[u8], max_px: u32) -> Result<DynamicImage, String> {
         .map_err(|e| format!("Failed to parse SVG: {e}"))?;
 
     let svg_size = tree.size;
-    let sw = svg_size.width() as f32;
-    let sh = svg_size.height() as f32;
+    let sw = svg_size.width();
+    let sh = svg_size.height();
 
     // Compute scale to fit within max_px × max_px
     let max_dim = if sw > sh { sw } else { sh };
